+++ agent/config.go
@@ -115,6 +115,9 @@ type Config struct {
 	// in check execution.
 	DisableAssets bool
 
+	// DisableSystemInfo stops the agent from collecting system information
+	DisableSystemInfo bool
+
 	// DisableSockets disables the event sockets
 	DisableSockets bool